tools/src/py/logminer/requirements.txt
jsonpickle==1.5.1
cachetools==4.2.1
pyformance==0.4
Werkzeug==0.16.1
future==0.18.2
pyyaml==5.4.1
absl-py==0.13.0
retrying==1.3.3
requests==2.26.0
drain3==0.9.6
confluent-kafka==1.7.0
kafka-python==2.0.2
PyMySQL==1.0.2
